refactor #13 : openfeign 컨벤션 맞춰 수정
